internal: continue types modularization
