+++ generators/base-application/support/prepare-entity.ts
@@ -27,7 +27,8 @@ import { getEntityParentPathAddition, getTypescriptKeyType } from '../../client/
 import { applicationTypes, databaseTypes, entityOptions, fieldTypes, searchEngineTypes } from '../../../lib/jhipster/index.js';
 import { binaryOptions } from '../../../lib/jdl/core/built-in-options/index.js';
 
-import type { Entity, PrimaryKey } from '../entity-all.js';
+import type { Entity } from '../entity-all.js';
+import type { PrimaryKey } from '../types.js';
 import type CoreGenerator from '../../base-core/generator.js';
 import type { ApplicationConfiguration } from '../application-config-all.js';
 import type { ApplicationAll } from '../application-properties-all.js';
@@ -270,7 +271,7 @@ export function derivedPrimaryKeyProperties(primaryKey: PrimaryKey) {
     typeString: primaryKey.type === STRING,
     typeLong: primaryKey.type === LONG,
     typeInteger: primaryKey.type === INTEGER,
-    typeNumeric: !primaryKey.composite && primaryKey.fields[0].fieldTypeNumeric,
+    typeNumeric: !primaryKey.composite && (primaryKey.fields[0] as any).fieldTypeNumeric,
   } as any);
 }
