+++ generators/client/support/template-utils.ts
@@ -20,7 +20,7 @@ import assert from 'node:assert';
 import path from 'node:path';
 
 import { clientFrameworkTypes, fieldTypes } from '../../../lib/jhipster/index.js';
-import type { PrimaryKey } from '../../base-application/entity-all.js';
+import type { PrimaryKey } from '../../base-application/types.js';
 import type { FieldType } from '../../base-application/internal/types/field-types.ts';
 import type { Field } from '../../base-application/field-all.js';
 import { getEntryIfTypeOrTypeAttribute } from './types-utils.js';
